refactor: activateClickButton을 home으로 이동
